KUDU-1113 pushed predicates not displaying on scans web page
This fixes Scanner so it will display predicates in /scanz. A Scanner
is constructed with a reference to its ScanSpec, but the predicates
in the ScanSpec are removed by the Scanner's iterator as they are
pushed down to lower iterators.

The Scanner now holds a copy of the ScanSpec, from just before the
ScanSpec is pass to the iterator and (potentially) modified.

Change-Id: I21821a1f7b5e4d5e5e1bdec17353ae5f3ef8e408
Reviewed-on: http://gerrit.cloudera.org:8080/4206
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>