Fix release notes formatting for multi-paragraph list items
If a list item spans multiple paragraphs in AsciiDoc, it needs
a '+' in the intervening blank line, and the latter paragraphs
should not be indented on their first line. The text is a bit
uglier, but it fixes the rendering.

Change-Id: Ie65570ed32acef1c3025f351f7c4b4944f443f6e
Reviewed-on: http://gerrit.cloudera.org:8080/4089
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>
(cherry picked from commit ffd8fa4758fd8598a8d06a79249a6ab57d72aa81)
Reviewed-on: http://gerrit.cloudera.org:8080/4090