KuduSession: do not advertise thread-safety
Do not advertise thread-safety of KuduSession methods for the
Kudu C++ client, even if they are thread-safe de facto.  This is
to have one set of semantics for both C++ and Java Kudu client
libraries (corresponding methods of the Java client are not
thread-safe).  Besides, current use cases for the Kudu client
assume that operations with KuduSession object do not involve
multiple threads.

Change-Id: I6b3fcbd0d6446ccddce7a1812260c692eba18fcd
Reviewed-on: http://gerrit.cloudera.org:8080/4105
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>