ksck: colorize and clean up output
Dan and I were looking at some ksck output earlier and found it somewhat
hard to read. This colorizes the output and also cleans up the format to
be slightly less messy. The output now goes to stdout instead of stderr
as well, so it's easier to pipe to a file or 'less', which is a common
use case.

Example output:

Tablet 0e4e79de2c2547c091779b13735d2b73 of table 'my_table' is under-replicated: 1 replica(s) not RUNNING
  63c4785c3b6f47f5a1c23ffbf6e52b71 (ts-023.foo.com:7050): RUNNING
  cd20c68cd23c4cf986eda0936a5691cc (ts-004.foo.com:7050): RUNNING [LEADER]
  5edf82f0516b4897b3a7991a7e67d71c (ts-028.foo.com:7050): bad state
    State:       NOT_STARTED
    Data state:  TABLET_DATA_COPYING
    Last status: TabletCopy: Downloading block 4611685629730158289 (3491/9569)

Change-Id: Icc5196d63cbcbcbb2a9aba1ff17377b678c104bd
Reviewed-on: http://gerrit.cloudera.org:8080/4129
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>