subprocess: use SIGKILL instead of SIGTERM death signal
This addresses a potential source of test timeouts if a test starts a
minicluster, pauses one of the daemons with SIGSTOP, and then exits
unexpectedly (eg due to a CHECK failure). With the SIGTERM death signal,
I found that this resulted in the test hanging on exit: apparently it
delivered the SIGTERM to the stopped child and was then waiting on it
indefinitely.

After changing to SIGKILL I see the broken test failing from the FATAL
as expected (rather than timing out).

Change-Id: Icc411b8c031c1c117bdf0d82f272da2263c97d88
Reviewed-on: http://gerrit.cloudera.org:8080/4139
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins