KUDU-687: expose additional tablet metadata in C++ client
This patch adds new data-only KuduReplica and KuduTablet classes.
Along with KuduTabletServer, the C++ client now has more parity with the
Java client w.r.t. exposing tablet metadata.

For now, this is only exposed via KuduScanToken, because it's already doing
the work to figure out which replicas exist where. That's an odd fit for
ksck (which doesn't scan, at least not using the client), but it should
stave off any controversy stemming from adding dubious public APIs. In the
future, it wouldn't be unreasonable for these classes to be exposed via
KuduTable in some way.

There are four public signature changes:
- Addition of KuduScanToken::tablet(): this is backwards compatible.
- Addition of KuduTabletServer::port(): this is backwards compatible.
- Change to the KuduScanToken constructor: this should be backwards
  compatible, because it's private and so shouldn't used by applications.
- Removal of KuduScanToken::TabletServers(): this is an incompatible change.
  I think it's OK because Impala is the only significant C++ client user and
  it's not even using scan tokens yet.

Change-Id: I3cbbb1df8d3adf60425541b57e68595bbf6e92ff
Reviewed-on: http://gerrit.cloudera.org:8080/4146
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>