compaction_policy: fix bound calculation
The upper-bound calculator for the knapsack problem maintains a heap
which is supposed to keep no more than one item beyond the maximum
weight allowed.

This invariant got broken in some cases when an item was added to the
heap -- in some cases a single addition actually requires removing
multiple old items.

A regression test for this fix is included in the following commit
("Optimize budgeted compaction policy with an approximation")

Change-Id: I76364c17debd7293b36884d64c7747f8c604ae12
Reviewed-on: http://gerrit.cloudera.org:8080/4152
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Tested-by: Kudu Jenkins