tracing: gzip the trace JSON ajax response
This enables gzipping of the actual trace which is sent back over HTTP
when using the tracing web page. I noticed that the traces can be quite
slow to receive, but given they often have a lot of repetitive text,
they are quite compressible.

The original Chrome implementation already did this, but I had ripped it
out for simplicity when importing the code. This simply adds it back on
the server side. A corresponding commit on the JS side adds a slight
tweak so that we can send back binary gzipped data, whereas the Chrome
version uses base64.

Change-Id: Ia8d5966e65b51576bf1460cb1ea8c2240ea88cbc
Reviewed-on: http://gerrit.cloudera.org:8080/4154
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>