transaction_tracker: back-off when logging in-flight transactions
On a test cluster pushed into overload, I ended up with thousands of
in-flight transactions which were only being drained at the rate of one
or two per second. Dumping the entire list of transactions once a second
ended up flooding the glog with many MB per second work of transaction
strings.

This changes the dumping to back-off exponentially.

Change-Id: I7d1a2563dd57c59ae8c130a5b26966f4964048de
Reviewed-on: http://gerrit.cloudera.org:8080/4159
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: David Ribeiro Alves <dralves@apache.org>