tool: port log-dump
This one was more complicated, because log-dump can run against a single
file or an entire tablet. So I put all the common functionality in one place
and referenced it from both modes.

I consolidated similar gflags where it made sense to do so, and I tweaked
the endline handling for help generation so that each parameter is separated
from the next with an empty line.

Semantic changes from log-dump:
- If called with print_entries=no, will also print the footer.
- The print_headers flag is now print_meta, to be more consistent with 'kudu
  cfile dump'.

Change-Id: I14f048169c0b211e3c72fcd255c76dd59cbb05c9
Reviewed-on: http://gerrit.cloudera.org:8080/4167
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>