KUDU-1586. consensus: always send at least one op
This fixes a bug in which the LogCache would not return any operations
in the case of a cache miss on an operation larger than the configured
batch size.

This would cause consensus to get "stuck": the leader would repeatedly
send status-only messages (i.e. no operations) to the follower in a
tight loop, never making any progress.

The new unit test reproduces the problem.

Change-Id: Ibde0e833a3bf02a5f09f2b73b6ab03b188c1e697
Reviewed-on: http://gerrit.cloudera.org:8080/4168
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>