jekyll: Use GitHub-flavored markdown and GitHub style highlighting
The default Kramdown syntax is different than GitHub's. Let's use
markdown syntax that is compatible with GitHub, which is what we had
intended to do all along. This allows us to use backticks for code
blocks, for example.

Also, instead of the default CodeRay syntax highlighter, use Rouge,
which looks cleaner.

Change-Id: I41257c97cbdbc471e427a116bdc58701892d42aa
Reviewed-on: http://gerrit.cloudera.org:8080/4189
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Mike Percy <mpercy@apache.org>