ts_tablet_manager: set status message after failed tablet copy
I noticed in a test cluster that a tablet copy had failed, and ksck was
reporting the state as:

  b314b10216d74c07ab9da974671137db (ve0136.halxg.cloudera.com:7050): bad state
    State:       NOT_STARTED
    Data state:  TABLET_DATA_TOMBSTONED
    Last status: TabletCopy: Downloading block 4611685870251245460 (2342/33956)

This was rather confusing, because it claimed it was downloading a
block, and also claimed to be tombstoned at the same time (rather than
being in the COPYING state).

This fixes the failure case to set the status message properly.

Change-Id: I36ecd45dacec4ea329a0f0dd236e52bc603f9b93
Reviewed-on: http://gerrit.cloudera.org:8080/4194
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>