maintenance_manager: simplify test
This simplifies maintenance_manager-test a bit to avoid a condition
variable, and instead just poll from the main thread. This reduces the
complexity to understand the test a bit, and the test still runs quickly
(<100ms).

This also changes the test operation in preparation for a later patch
which will allow an operation to run multiple times in parallel and
auto-disable itself after a prescribed number of runs.

Change-Id: I265318f64771803714f48a6e96fd3260b6eafae4
Reviewed-on: http://gerrit.cloudera.org:8080/4294
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins