KUDU-1495. Maintenance manager should not schedule new ops during unregister
This fixes a bug where deleting a tablet could block for an extended
time on compactions being scheduled if the maintenance manager was
configured with multiple threads.

The issue was that Unregister() waited for the number of running
operations to reach 0 before removing the op from the candidate list of
ops, but did not prevent new instances of that op from being scheduled.

Change-Id: I3675705caf5b73f8a480036b974e4db6c205616a
Reviewed-on: http://gerrit.cloudera.org:8080/4295
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins