[server] clean-up: stringstream --> ostringstream
Replaced std::stringstream with std::ostringstream in places
where only output stream functionality is required from the
stream buffer class.

Include iosfwd instead of stream headers for forward declarations.
Minor clean-up on header files include order to be in line
with the styling guide.

There are not any functional changes in this changelist.

Change-Id: I059c19bb287ee37f96e5ad42f4886015a1697d19
Reviewed-on: http://gerrit.cloudera.org:8080/4303
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>