KUDU-861 Support changing default, storage attributes
This patch adds support for adding, changing, or removing column defaults
and changing the storage attributes of a column. Changes to a column are
encoded as a ColumnSchemaDelta, which can be merged with a ColumnSchema
to change it.

Changing type and nullability of a column is still unsupported.

No failures in 100+ iterations of alter_table-randomized-test.

I also ran into an issue with altering RLE columns:
1. Add an RLE-encoded column to a table
2. Alter the column
3. Scan the column
3 will cause a check failure on scanning to pos 0 of an empty RLE
block. Test and fix included.

Change-Id: I457d99ba2188ef6e439df47c0d94f2dc1a62ea6c
Reviewed-on: http://gerrit.cloudera.org:8080/4310
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>