KUDU-1593 - [python] Allow to set the number of replicas per tablet at table creation time
Currently the Kudu Python client does not allow for the number of replicas to be set at
table creation time because the num_replicas method of the KuduTableCreator is not
exposed to the Python class. This patch exposes that method as well as it exposes the
num_replicas method of the KuduTable class so that the number of replicas for a table
can be retrieved from the cluster.  This patch includes a test.

Change-Id: I9dcac887029ef22e850ea8be72e71b684e6bee05
Reviewed-on: http://gerrit.cloudera.org:8080/4315
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Tested-by: Kudu Jenkins