log_util: downgrade scary "missing footer" message to INFO
In most cases if a Kudu server restarts, it will leave log segments with
no footer. This previously caused a scary-looking warning to be emitted
on the next startup.

This downgrades the message to INFO for the case of the footer not being
found. I also fixed this code path to properly propagate other types of
errors (eg IOError) which might occur while reading the footer.

Change-Id: Ic60f57e1276e6e4cc209d37d6fdad444c24db3c6
Reviewed-on: http://gerrit.cloudera.org:8080/4319
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>