tablet_server-stress-test: limit runtime of this test
This stress test is typically the long pole in the dist-test runtime.
For example, in a recent ASAN run[1] the test took around 730 seconds.

This changes the behavior of the test to run for a prescribed amount of
time (60 seconds in slow mode, 10 in fast). This should keep relatively
good coverage while avoiding such long test runs.

[1] http://dist-test.cloudera.org/trace?job_id=jenkins-slave.1473295331.9755

Change-Id: I7441f50bcd4788e3e54a90bd5f782201a7d4c6af
Reviewed-on: http://gerrit.cloudera.org:8080/4329
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins