KUDU-1135 (part 1): avoid flushing cmeta to disk twice when voting
This adds a flag to the term advancement calls which prevents it from
flushing to disk. We use this flag during voting when we know we are
about to flush our vote to disk immediately following the term
advancement.

This is part 1: another patch will do the same optimization for the case
of starting an election.

Change-Id: Iecc55bc9e96dcdc918ede1190b7cbac719f95506
Reviewed-on: http://gerrit.cloudera.org:8080/4333
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Tested-by: David Ribeiro Alves <dralves@apache.org>