Fix flakiness in tablet_replacement-itest
This test was flaky because we were trying to perform a config change
when it was possible that the leader had not yet committed an operation
in its term.

Without the fix, I looped it 100 times with --stress_cpu_threads=8 and
it failed twice due to this issue. With the fix, I looped it 400 times
with the stress threads and only saw one failure, which seemed like a
straight timeout due to the stress threads monopolizing the CPU.

Change-Id: I15b361179f5e20f848f5b346e202b9217a35b61d
Reviewed-on: http://gerrit.cloudera.org:8080/4342
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>