master: do not crash on /dump-entities during startup
A user reported a master crash when the following conditions were met:
1. The master tablet was enormous and took a while to bootstrap.
2. An external monitoring agent was periodically accessing the master's
   /dump-entities URL.

While fixing this, I also double checked that the rest of the master URLs do
the right thing before making calls into the master.

The majority of the patch repurposes the PeriodicWebUIChecker (from
linked_list-test) for the new integration test. I modified its URL polling
behavior so that all URLs are polled before sleeping.

Change-Id: I7d8e9caf7236d2c487fb9322cb3e1993cc631cdf
Reviewed-on: http://gerrit.cloudera.org:8080/4366
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>