tablet_history_gc-itest: set scanner_ttl_ms high
This test is currently flaky on slower builds (TSAN/ASAN) because of
scanner expiration. The test opens scanners and then waits some time
before actually reading from them. In a slow build, this might be more
than the default 1-minute scanner TTL, which would cause the scanner to
expire and the test to fail to read.

Change-Id: I4bd8a5bdc77871a95b3630f1622a802967d14b84
Reviewed-on: http://gerrit.cloudera.org:8080/4375
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>