flex_partitioning-itest: shard into separate cases
Rather than doing a loop in the main test case, shard the test into one
test case per setup.

This slightly increases the runtime of a non-parallel invocation, since
we now build and tear down a cluster for each configuration. However,
this permits parallelizing its run on dist-test. Since this is one of
the longer-running tests, it's worth it.

Change-Id: I6d352be66573fb95aac1e3ee8bc74569d4e8a0b3
Reviewed-on: http://gerrit.cloudera.org:8080/4376
Reviewed-by: Dan Burkert <dan@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>