KUDU-1604 - [python] Fix bug getting table column by index
There is currently a bug when attempting to access a table column
by its index. When attempted, the column definition returned has
the index value as the column name. This patch fixes this bug so
that the appropriate column name is set and returned. An existing
test has been updated to account for this usage.

Change-Id: If754ddb463223430c7fe3c8920e433031162562e
Reviewed-on: http://gerrit.cloudera.org:8080/4378
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>