Add script to run clang-tidy against a patch
This tooling runs clang-tidy against a given patch, and formats the
resulting warnings in a form that can be consumed by the gerrit API[1]

This has been running out of a private branch for many months and
proven generally useful. This patch also updates .clang-tidy based on
the configuration that has been running.

[1] https://gerrit-review.googlesource.com/Documentation/rest-api-changes.html#set-review

Change-Id: I6c6246f4691b0ca143eea4aa8fb6a4b23099ce4e
Reviewed-on: http://gerrit.cloudera.org:8080/4381
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins